flask_app/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
click==7.1.2
    # via flask
flask==1.1.4
    # via -r requirements.in
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via jinja2
psycopg2-binary==2.8.6
    # via -r requirements.in
simplejson==3.17.6
    # via -r requirements.in
werkzeug==1.0.1
    # via flask
